feat: dynamic provider config and controller refactor
- Generate openclaw.json providers from Claw CR credentials with Provider field
- Split controller into focused files (proxy, status, credentials)

Made-with: Cursor
