+++ README.md
@@ -304,8 +304,7 @@ oc exec -n claw-operator deployment/claw -- \
   node /app/dist/index.js devices list
 
 # Approve by request ID (copy from the Pending table above)
-oc exec -n claw-operator deployment/claw -- \
-  node /app/dist/index.js devices approve <requestId>
+<requestId>
 ```
 
 Refresh the browser after approval. The device is remembered -- you won't need to pair again unless you clear browser data or switch browsers.
